devtools/server/tests/mochitest/test_inspector_getImageData-wait-for-load.html

<!DOCTYPE HTML>
<html>
<!--
Tests for InspectorActor.getImageData() in following cases:
 * Image takes too long to load (the method rejects after a timeout).
 * Image is loading when the method is called and the load finishes before
   timeout.
 * Image fails to load.

https://bugzilla.mozilla.org/show_bug.cgi?id=1192536
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1192536</title>

  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript" src="inspector-helpers.js"></script>
  <script type="application/javascript">
"use strict";

const PATH = "http://mochi.test:8888/chrome/devtools/server/tests/mochitest/";
const BASE_IMAGE = PATH + "inspector-delay-image-response.sjs";
const DELAYED_IMAGE = BASE_IMAGE + "?delay=300";
const TIMEOUT_IMAGE = BASE_IMAGE + "?delay=50000";
const NONEXISTENT_IMAGE = PATH + "this-does-not-exist.png";

window.onload = function() {
  SimpleTest.waitForExplicitFinish();
  runNextTest();
};

function pushPref(preferenceName, value) {
  return new Promise(resolve => {
    const options = {"set": [[preferenceName, value]]};
    SpecialPowers.pushPrefEnv(options, resolve);
  });
}

let gImg = null;
let gNodeFront = null;
let gWalker = null;

addTest(function setup() {
  const url = document.getElementById("inspectorContent").href;
  attachURL(url, function(err, client, tab, doc) {
    const {InspectorFront} = require("devtools/shared/fronts/inspector");
    const inspector = InspectorFront(client, tab);

    promiseDone(inspector.getWalker().then(walker => {
      gWalker = walker;
      // eslint-disable-next-line max-nested-callbacks
      return walker.querySelector(gWalker.rootNode, "img.custom").then(img => {
        gNodeFront = img;
        gImg = doc.querySelector("img.custom");

        ok(gNodeFront, "Got the image NodeFront.");
        ok(gImg, "Got the image Node.");
      });
    }).then(runNextTest));
  });
});

addTest(async function testTimeout() {
  info("Testing that the method aborts if the image takes too long to load.");

  // imageToImageData() only times out when flags.testing is not set.
  await pushPref("devtools.testing", false);

  gImg.src = TIMEOUT_IMAGE;

  info("Calling getImageData().");
  ensureRejects(gNodeFront.getImageData(), "Timeout image").then(runNextTest);
});

addTest(async function testNonExistentImage() {
  info("Testing that non-existent image causes a rejection.");

  // This test shouldn't hit the timeout.
  await pushPref("devtools.testing", true);

  gImg.src = NONEXISTENT_IMAGE;

  info("Calling getImageData().");
  ensureRejects(gNodeFront.getImageData(), "Non-existent image").then(runNextTest);
});

addTest(async function testDelayedImage() {
  info("Testing that the method waits for an image to load.");

  // This test shouldn't hit the timeout.
  await pushPref("devtools.testing", true);

  gImg.src = DELAYED_IMAGE;

  info("Calling getImageData().");
  checkImageData(gNodeFront.getImageData()).then(runNextTest);
});

addTest(function cleanup() {
  gImg = null;
  gNodeFront = null;
  gWalker = null;
  runNextTest();
});

/**
 * Asserts that the given promise rejects.
 */
function ensureRejects(promise, desc) {
  return promise.then(() => {
    ok(false, desc + ": promise resolved unexpectedly.");
  }, () => {
    ok(true, desc + ": promise rejected as expected.");
  });
}

/**
 * Waits for the call to getImageData() the resolve and checks that the image
 * size is reported correctly.
 */
function checkImageData(promise, { width, height } = { width: 1, height: 1 }) {
  return promise.then(({ size }) => {
    is(size.naturalWidth, width, "The width is correct.");
    is(size.naturalHeight, height, "The height is correct.");
  });
}

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1192536">Mozilla Bug 1192536</a>
<a id="inspectorContent" target="_blank" href="inspector_getImageData.html">Test Document</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
